Ensure ZNC can bind to ports < 1024
To serve the webadmin over HTTPS, we need to bind to port 443, but
that's restricted to root users by default. To allow us to run our
service as non-root (for various reasons), we should permit ZNC to use
the `net_bind_service` capability.

This also adds a little more detail to the README.