Install certbot to manage SSL certificates
Managing [Let's Encrypt](https://letsencrypt.org/) SSL certificates is
best done with [certbot](https://certbot.eff.org). To install on Linux,
it's best to use the package archive. Because we don't want Make to
compare `certbot`'s last-update time to the PPA's, the PPA is an
"order-only" dependency, as denoted by `|`.